requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
streamlit>=1.24.0
pandas>=2.0.0
shiny>=0.6.0
faicons>=0.2.0
openpyxl>=3.1.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
